+++ synquill/example/lib/models/todo.dart
@@ -27,7 +27,10 @@ mixin TodoApiAdapter on BasicApiAdapter<Todo> {
       baseUrl.resolve('$pluralType/$id');
 
   @override
-  FutureOr<Uri> urlForFindAll({Map<String, dynamic>? extra}) async {
+  FutureOr<Uri> urlForFindAll({
+    QueryParams? queryParams,
+    Map<String, dynamic>? extra,
+  }) async {
     final users = await SynquillStorage.instance.users
         .findAll(loadPolicy: DataLoadPolicy.localOnly);